Refactor import paths for UI components
