+++ src/components/ui/Em/Em.stories.js
@@ -1,5 +1,6 @@
 import Em from './Em';
-import {Heading} from '~/';
+import Heading from '../Heading/Heading';
+
 import SandboxEditor from '~/components/tools/SandboxEditor/SandboxEditor';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
