+++ src/components/ui/Link/Link.stories.js
@@ -1,5 +1,4 @@
 import Link from './Link';
-import {Text} from '~/';
 import SandboxEditor from '~/components/tools/SandboxEditor/SandboxEditor';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
