--- src/index.js
@@ -1,32 +1,0 @@
-'use client';
-
-// export Button as default
-export {default as Button} from './components/ui/Button/Button.js';
-
-export {default as Tabs} from './components/ui/Tabs/Tabs.js';
-
-export {default as Separator} from './components/ui/Separator/Separator.js';
-
-export {default as Badge} from './components/ui/Badge/Badge.js';
-
-export {default as Em} from './components/ui/Em/Em.js';
-
-export {default as Heading} from './components/ui/Heading/Heading.js';
-
-export {default as Text} from './components/ui/Text/Text.js';
-
-export {default as BlockQuote} from './components/ui/BlockQuote/BlockQuote.js';
-
-export {default as Link} from './components/ui/Link/Link.js';
-
-export {default as Quote} from './components/ui/Quote/Quote.js';
-
-export {default as Code} from './components/ui/Code/Code.js';
-
-export {default as Avatar} from './components/ui/Avatar/Avatar.tsx';
-
-export {default as Callout} from './components/ui/Callout/Callout.js';
-
-export {default as Dropdown} from './components/ui/Dropdown/Dropdown.js';
-
-export {default as Tooltip} from './components/ui/Tooltip/Tooltip.js';
